+++ CHANGES.md
@@ -1,5 +1,7 @@
-# 4.1.2 (2019-07-29)
-- [FIXED] Plugins can now be loaded from outside of the 'plugins/' directory
+# UNRELEASED
+- [NEW] Added option to set new IAM API key.
+- [FIXED] Allow plugins to be loaded from outside the 'plugins/' directory.
+- [FIXED] Retry bad IAM token requests.
 
 # 4.1.1 (2019-06-17)
 - [FIXED] Remove unnecessary `npm-cli-login` dependency.